Fix scanThreshold setting not working

+++ HashCalculator/CmdOptions.cs
@@ -12,11 +12,9 @@ namespace HashCalculator
         public int ScanInterval { get; set; }
 
         [Option('t', "threshold", Default = 20, Required = false, HelpText = "Scan threshold (Gigabyte)")]
-        public int ScanThreshold { get; set; }
+        public long ScanThreshold { get; set; }
 
         [Option('v', "verbose", Required = false, HelpText = "Print verbose message")]
         public bool IsVerbose { get; set; }     
-
-
     }
 }
